Added tests for src\client\interpreter\configuration\pythonPathUpdaterServiceFactory.ts
